Speed up reading with `xdecimate` option for small decimations
When reading files with small values of the distance decimation,
the HDF5 library appears to make lots of individual reads from the file,
making it much slower than reading the whole chunk in one go and then
decimating in memory.  Use this method when the decimation value is
small.

Note that this requires that the whole undecimated time span is read
into memory.