Fix reading of v2 files
A1 files with version greater than v2.3.13 change the reference
point of the block date to be the middle of the block, so update
the block start dates to account for this.

Such files also allow for strain or strain rate to be present,
so accommodate this as well.  For now, always return strain if
present, then strain rate.